+++ hcl2template/utils.go
@@ -1,11 +1,13 @@
 package hcl2template
 
 import (
+	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
 
+	"github.com/gobwas/glob"
 	"github.com/hashicorp/hcl/v2"
 )
 
@@ -87,4 +89,23 @@ func GetHCL2Files(filename, hclSuffix, jsonSuffix string) (hclFiles, jsonFiles [
 	}
 
 	return hclFiles, jsonFiles, diags
+}
+
+// Convert -only and -except globs to glob.Glob instances.
+func convertFilterOption(patterns []string, optionName string) ([]glob.Glob, hcl.Diagnostics) {
+	var globs []glob.Glob
+	var diags hcl.Diagnostics
+
+	for _, pattern := range patterns {
+		g, err := glob.Compile(pattern)
+		if err != nil {
+			diags = append(diags, &hcl.Diagnostic{
+				Summary:  fmt.Sprintf("Invalid -%s pattern %s: %s", optionName, pattern, err),
+				Severity: hcl.DiagError,
+			})
+		}
+		globs = append(globs, g)
+	}
+
+	return globs, diags
 }
